+++ jdbc/src/main/java/nextstep/jdbc/support/ObjectUtils.java
@@ -1,0 +1,10 @@
+package nextstep.jdbc.support;
+
+public class ObjectUtils {
+
+    public static <T> void notNull(final T object, final String name) {
+        if (object == null) {
+            throw new IllegalArgumentException(name + " must be not null.");
+        }
+    }
+}
