+++ jdbc/src/test/java/nextstep/jdbc/support/StringUtilsTest.java
@@ -1,22 +1,12 @@
 package nextstep.jdbc.support;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ValidUtilsTest {
-
-    @DisplayName("객체가 null일 경우 예외를 던진다.")
-    @Test
-    void notNull() {
-        // given
-        final Object object = null;
-
-        // when & then
-        assertThatThrownBy(() -> ValidUtils.notNull(object, "Object"))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
+class StringUtilsTest {
 
     @DisplayName("문자열이 blank일 경우 예외를 던진다.")
     @Test
@@ -25,7 +15,7 @@ class ValidUtilsTest {
         final var str = " ";
 
         // when & then
-        assertThatThrownBy(() -> ValidUtils.notBlank(str, "SQL"))
+        assertThatThrownBy(() -> StringUtils.notBlank(str, "SQL"))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
